ethereum_standards::IERC20

Enum IERC20Calls

pub enum IERC20Calls {
    totalSupply(totalSupplyCall),
    balanceOf(balanceOfCall),
    transfer(transferCall),
    allowance(allowanceCall),
    approve(approveCall),
    transferFrom(transferFromCall),
}

Container for all the IERC20 function calls.

Variants

totalSupply(totalSupplyCall)

balanceOf(balanceOfCall)

transfer(transferCall)

allowance(allowanceCall)

approve(approveCall)

transferFrom(transferFromCall)

Implementations

impl IERC20Calls

pub const SELECTORS: &'static [[u8; 4]]

All the selectors of this enum.

Note that the selectors might not be in the same order as the variants. No guarantees are made about the order of the selectors.

Prefer using SolInterface methods instead.