ethereum_standards::IERC20

Struct transferCall

pub struct transferCall {
    pub to: Address,
    pub value: U256,
}

@dev Moves a value amount of tokens from the caller’s account to to.

Returns a boolean value indicating whether the operation succeeded.

Emits a {Transfer} event. Function with signature transfer(address,uint256) and selector 0xa9059cbb.

function transfer(address to, uint256 value) external returns (bool);

Fields
to: Addressvalue: U256